Extract `cvd/fetch/get_optional.h`
Bug: b/424021019
Test: bazel test '//...'
