+++ base/cvd/cuttlefish/host/commands/cvd/fetch/BUILD.bazel
@@ -144,6 +144,7 @@ cc_library(
         "//cuttlefish/host/commands/cvd/fetch:extract_image_contents",
         "//cuttlefish/host/commands/cvd/fetch:fetch_cvd_parser",
         "//cuttlefish/host/commands/cvd/fetch:fetch_tracer",
+        "//cuttlefish/host/commands/cvd/fetch:get_optional",
         "//cuttlefish/host/commands/cvd/fetch:host_package",
         "//cuttlefish/host/commands/cvd/fetch:substitute",
         "//cuttlefish/host/commands/cvd/fetch:vector_flags",
@@ -216,6 +217,18 @@ cc_library(
 clang_tidy_test(
     name = "fetch_tracer_clang_tidy",
     srcs = [":fetch_tracer"],
+    tags = ["clang_tidy", "clang-tidy"],
+)
+
+cc_library(
+    name = "get_optional",
+    hdrs = ["get_optional.h"],
+    copts = COPTS,
+)
+
+clang_tidy_test(
+    name = "get_optional_clang_tidy",
+    srcs = [":get_optional"],
     tags = ["clang_tidy", "clang-tidy"],
 )
