+++ base/cvd/cuttlefish/host/commands/cvd/fetch/fetch_cvd.cc
@@ -38,6 +38,7 @@
 #include "cuttlefish/host/commands/cvd/fetch/extract_image_contents.h"
 #include "cuttlefish/host/commands/cvd/fetch/fetch_cvd_parser.h"
 #include "cuttlefish/host/commands/cvd/fetch/fetch_tracer.h"
+#include "cuttlefish/host/commands/cvd/fetch/get_optional.h"
 #include "cuttlefish/host/commands/cvd/fetch/host_package.h"
 #include "cuttlefish/host/commands/cvd/fetch/vector_flags.h"
 #include "cuttlefish/host/libs/config/fetcher_config.h"
@@ -107,17 +108,6 @@ struct HostToolsTarget {
 bool ShouldAppendSubdirectory(const FetchFlags& flags) {
   return flags.vector_flags.NumberOfBuilds().value_or(1) > 1 ||
          !flags.vector_flags.target_subdirectory.empty();
-}
-
-template <typename T>
-std::optional<T> GetOptional(const std::vector<T>& vector, size_t i) {
-  return i < vector.size() ? std::optional(vector[i]) : std::nullopt;
-}
-
-template <typename T>
-std::optional<T> GetOptional(const std::vector<std::optional<T>>& vector,
-                             size_t i) {
-  return i < vector.size() ? vector[i] : std::nullopt;
 }
 
 BuildStrings GetBuildStrings(const VectorFlags& flags, const int index) {
